Fix harvester dock sequence names
