+++ OpenRA.Mods.RA/OpenRA.Mods.RA.csproj
@@ -80,7 +80,7 @@
     <Compile Include="Activities\Infiltrate.cs" />
     <Compile Include="Activities\LayMines.cs" />
     <Compile Include="Activities\Leap.cs" />
-    <Compile Include="Activities\RAHarvesterDockSequence.cs" />
+    <Compile Include="Activities\HarvesterDockSequence.cs" />
     <Compile Include="Activities\Teleport.cs" />
     <Compile Include="AI\SupportPowerDecision.cs" />
     <Compile Include="Effects\GpsSatellite.cs" />
